# nodejs-server/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is the API for the Estuary application."
  version: "0.0.0"
  title: "Estuary API"
  termsOfService: "http://estuary.tech"
  contact:
    name: "API Support"
    url: "https://docs.estuary.tech/feedback"
  license:
    name: "Apache 2.0 Apache-2.0 OR MIT"
    url: "https://github.com/application-research/estuary/blob/master/LICENSE.md"
host: "api.estuary.tech"
basePath: "/"
security:
- bearerAuth: []
paths:
...
  /collections/{coluuid}/contents:
    delete:
      tags:
      - "collections"
      summary: "Deletes a content from a collection"
      description: "This endpoint is used to delete an existing content from an existing\
        \ collection. If two or more files with the same contentid exist in the collection,\
        \ delete the one in the specified path"
      operationId: "collectionsColuuidContentsDELETE"
      produces:
      - "application/json"
      parameters:
      - name: "coluuid"
        in: "path"
        description: "Collection ID"
        required: true
        type: "string"
      - name: "contentid"
        in: "path"
        description: "Content ID"
        required: true
        type: "string"
      - in: "body"
        name: "by"
        description: "Variable to use when filtering for files (must be either 'path'\
          \ or 'content_id')"
        required: true
        schema:
          type: "string"
      - in: "body"
        name: "value"
        description: "Value of content_id or path to look for"
        required: true
        schema:
          type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            type: "string"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/util.HttpError"
      x-swagger-router-controller: "Collections"
  /content/add:
    post:
      tags:
      - "content"
      summary: "Add new content"
      description: "This endpoint is used to upload new content."
      operationId: "contentAddPOST"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - name: "data"
        in: "formData"
        description: "File to upload"
        required: true
        type: "file"
      - name: "coluuid"
        in: "path"
        description: "Collection UUID"
        required: true
        type: "string"
      - name: "dir"
        in: "path"
        description: "Directory"
        required: true
        type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/util.ContentAddResponse"
      x-swagger-router-controller: "Content"
  /content/add-car:
    post:
      tags:
      - "content"
      summary: "Add Car object"
      description: "This endpoint is used to add a car object to the network. The\
        \ object can be a file or a directory."
      operationId: "contentAdd_carPOST"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Car"
        required: true
        schema:
          type: "string"
      - name: "filename"
        in: "query"
        description: "Filename"
        required: false
        type: "string"
      - name: "commp"
        in: "query"
        description: "Commp"
        required: false
        type: "string"
      - name: "size"
        in: "query"
        description: "Size"
        required: false
        type: "string"
      responses: {}
      x-swagger-router-controller: "Content"
  /content/add-ipfs:
    post:
      tags:
      - "content"
      summary: "Add IPFS object"
      description: "This endpoint is used to add an IPFS object to the network. The\
        \ object can be a file or a directory."
      operationId: "contentAdd_ipfsPOST"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "IPFS Body"
        required: true
        schema:
          $ref: "#/definitions/util.ContentAddIpfsBody"
      responses: {}
      x-swagger-router-controller: "Content"
  /content/aggregated/{content}:
    get:
      tags:
      - "content"
      summary: "Get aggregated content stats"
      description: "This endpoint returns aggregated content stats"
      operationId: "contentAggregatedContentGET"
      produces:
      - "application/json"
      parameters:
      - name: "content"
        in: "path"
        description: "Content ID"
        required: true
        type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            type: "string"
      x-swagger-router-controller: "Content"
  /content/all-deals:
    get:
      tags:
      - "content"
      summary: "Get all deals for a user"
      description: "This endpoint is used to get all deals for a user"
      operationId: "contentAll_dealsGET"
      produces:
      - "application/json"
      parameters:
      - name: "begin"
        in: "query"
        description: "Begin"
        required: true
        type: "string"
      - name: "duration"
        in: "query"
        description: "Duration"
        required: true
        type: "string"
      - name: "all"
        in: "query"
        description: "All"
        required: true
        type: "string"
      responses: {}
      x-swagger-router-controller: "Content"
  /content/bw-usage/{content}:
    get:
      tags:
      - "content"
      summary: "Get content bandwidth"
      description: "This endpoint returns content bandwidth"
      operationId: "contentBw_usageContentGET"
      produces:
      - "application/json"
      parameters:
      - name: "content"
        in: "path"
        description: "Content ID"
        required: true
        type: "string"
      responses: {}
      x-swagger-router-controller: "Content"
  /content/create:
    post:
      tags:
      - "content"
      summary: "Add a new content"
      description: "This endpoint adds a new content"
      operationId: "contentCreatePOST"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Content"
        required: true
        schema:
          type: "string"
      responses: {}
      x-swagger-router-controller: "Content"
  /content/deals:
    get:
      tags:
      - "content"
      summary: "Content with deals"
      description: "This endpoint lists all content with deals"
      operationId: "contentDealsGET"
      produces:
      - "application/json"
      parameters:
      - name: "limit"
        in: "query"
        description: "Limit"
        required: false
        type: "integer"
      - name: "offset"
        in: "query"
        description: "Offset"
        required: false
        type: "integer"
      responses: {}
      x-swagger-router-controller: "Content"
  /content/ensure-replication/{datacid}:
    get:
      tags:
      - "content"
      summary: "Ensure Replication"
      description: "This endpoint ensures that the content is replicated to the specified\
        \ number of providers"
      operationId: "contentEnsure_replicationDatacidGET"
      produces:
      - "application/json"
      parameters:
      - name: "datacid"
        in: "path"
        description: "Data CID"
        required: true
        type: "string"
      responses: {}
      x-swagger-router-controller: "Content"
  /content/failures/{content}:
    get:
      tags:
      - "content"
      summary: "List all failures for a content"
      description: "This endpoint returns all failures for a content"
      operationId: "contentFailuresContentGET"
      produces:
      - "application/json"
      parameters:
      - name: "content"
        in: "path"
        description: "Content ID"
        required: true
        type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            type: "string"
      x-swagger-router-controller: "Content"
  /content/importdeal:
    post:
      tags:
      - "content"
      summary: "Import a deal"
      description: "This endpoint imports a deal into the shuttle."
      operationId: "contentImportdealPOST"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Import a deal"
        required: true
        schema:
          $ref: "#/definitions/main.importDealBody"
      responses: {}
      x-swagger-router-controller: "Content"
...
  /content/read/{cont}:
    get:
      tags:
      - "content"
      summary: "Read content"
      description: "This endpoint reads content from the blockstore"
      operationId: "contentReadContGET"
      produces:
      - "application/json"
      parameters:
      - name: "cont"
        in: "path"
        description: "CID"
        required: true
        type: "string"
      responses: {}
      x-swagger-router-controller: "Content"
...
  /content/stats:
    get:
      tags:
      - "content"
      summary: "Get content statistics"
      description: "This endpoint is used to get content statistics. Every content\
        \ stored in the network (estuary) is tracked by a unique ID which can be used\
        \ to get information about the content. This endpoint will allow the consumer\
        \ to get the collected stats of a conten"
      operationId: "contentStatsGET"
      produces:
      - "application/json"
      parameters:
      - name: "limit"
        in: "path"
        description: "limit"
        required: true
        type: "string"
      responses: {}
      x-swagger-router-controller: "Content"
  /content/status/{id}:
    get:
      tags:
      - "content"
      summary: "Content Status"
      description: "This endpoint returns the status of a content"
      operationId: "contentStatusIdGET"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Content ID"
        required: true
        type: "integer"
      responses: {}
      x-swagger-router-controller: "Content"
...
securityDefinitions:
  bearerAuth:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
...
  main.importDealBody:
    type: "object"
    properties:
      coluuid:
        type: "string"
      dealIDs:
        type: "array"
        items:
          type: "integer"
      dir:
        type: "string"
      name:
        type: "string"
...
  util.ContentAddIpfsBody:
    type: "object"
    properties:
      coluuid:
        type: "string"
      dir:
        type: "string"
      filename:
        type: "string"
      peers:
        type: "array"
        items:
          type: "string"
      root:
        type: "string"
  util.ContentAddResponse:
    type: "object"
    properties:
      cid:
        type: "string"
      estuaryId:
        type: "integer"
      providers:
        type: "array"
        items:
          type: "string"
      retrieval_url:
        type: "string"
    example:
      retrieval_url: "retrieval_url"
      estuaryId: 0
      providers:
      - "providers"
      - "providers"
      cid: "cid"
  util.HttpError:
    type: "object"
    properties:
      code:
        type: "integer"
      details:
        type: "string"
      reason:
        type: "string"